moved another function, and tests, so much tests... moving to api
