+++ src/test/java/com/group16b/ApplicationLayer/PurchasePolicyServiceTests.java
@@ -1,9 +1,6 @@
 package com.group16b.ApplicationLayer;
 
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -12,8 +9,6 @@ import java.time.LocalDateTime;
 import java.util.Map;
 import java.util.TreeMap;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.atomic.AtomicReference;
 
 import com.group16b.ApplicationLayer.Records.PurchasePolicyRecord;
 import org.junit.jupiter.api.BeforeEach;
@@ -25,7 +20,6 @@ import com.group16b.ApplicationLayer.Records.EventRecord;
 import com.group16b.DomainLayer.Event.Event;
 import com.group16b.DomainLayer.Event.IEventRepository;
 import com.group16b.DomainLayer.Interfaces.IRepository;
-import com.group16b.DomainLayer.Policies.PurchasePolicy.LotteryPolicy;
 import com.group16b.DomainLayer.ProductionCompany.IProductionCompanyRepository;
 import com.group16b.DomainLayer.ProductionCompany.ProductionCompany;
 import com.group16b.DomainLayer.User.User;
@@ -57,8 +51,6 @@ public class PurchasePolicyServiceTests {
     private Location location1;
     private Segment segment1;
     private ProductionCompany company;
-
-    private LocalDateTime now = LocalDateTime.now();
 
     @BeforeEach
     public void setUp() {
@@ -109,236 +101,6 @@ public class PurchasePolicyServiceTests {
         eventRepository.save(e1);
         venue1.bookEvent(e1.getEventStartTime(), e1.getEventEndTime(), 1);
         venueRepository.save(venue1);
-    }
-
-    
-
-    @Test
-    public void enrollLotteryPolicy_Success() {
-        Event e = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        LotteryPolicy lotteryPolicy = new LotteryPolicy(0, "Lottery", 50, now.plusDays(5));
-        e.setLotteryPolicy(lotteryPolicy);
-        eventRepository.save(e);
-        Result<Boolean> res = purchasePolicyService.enrollInLottery("user1", e1.getEventID());
-        assertTrue(res.isSuccess());
-        Event e2 = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertTrue(e2.getLotteryPolicy().getParticipants().contains("testuser"));
-    }
-
-    @Test
-    public void enrollLotteryPolicy_FailInvalidToken() {
-        Event e = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        LotteryPolicy lotteryPolicy = new LotteryPolicy(0, "Lottery", 50, now.plusDays(5));
-        e.setLotteryPolicy(lotteryPolicy);
-        eventRepository.save(e);
-        Result<Boolean> res = purchasePolicyService.enrollInLottery("invalid_token", e1.getEventID());
-        assertFalse(res.isSuccess());
-        Event e2 = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertFalse(e2.getLotteryPolicy().getParticipants().contains("testuser"));
-    }
-
-    @Test
-    public void enrollLotteryPolicy_FailNotUserToken() {
-        Event e = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        LotteryPolicy lotteryPolicy = new LotteryPolicy(0, "Lottery", 50, now.plusDays(5));
-        e.setLotteryPolicy(lotteryPolicy);
-        eventRepository.save(e);
-        Result<Boolean> res = purchasePolicyService.enrollInLottery("guest", e1.getEventID());
-        assertFalse(res.isSuccess());
-        Event e2 = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertFalse(e2.getLotteryPolicy().getParticipants().contains("testuser"));
-    }
-
-    @Test
-    public void enrollLotteryPolicy_FailUserNotFound() {
-        Event e = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        LotteryPolicy lotteryPolicy = new LotteryPolicy(0, "Lottery", 50, now.plusDays(5));
-        e.setLotteryPolicy(lotteryPolicy);
-        eventRepository.save(e);
-        userRepository.delete("testuser");
-        Result<Boolean> res = purchasePolicyService.enrollInLottery("user1", e1.getEventID());
-        assertFalse(res.isSuccess());
-        Event e2 = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertFalse(e2.getLotteryPolicy().getParticipants().contains("testuser"));
-    }
-
-    @Test
-    public void enrollLotteryPolicy_FailEventNotFound() {
-        Event e = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        LotteryPolicy lotteryPolicy = new LotteryPolicy(0, "Lottery", 50, now.plusDays(5));
-        e.setLotteryPolicy(lotteryPolicy);
-        eventRepository.save(e);
-        Result<Boolean> res = purchasePolicyService.enrollInLottery("user1", 999);
-        assertFalse(res.isSuccess());
-        Event e2 = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertFalse(e2.getLotteryPolicy().getParticipants().contains("testuser"));
-    }
-
-    @Test
-    public void enrollLotteryPolicy_FailInactiveEvent() {
-        Event e = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        LotteryPolicy lotteryPolicy = new LotteryPolicy(0, "Lottery", 50, now.plusDays(5));
-        e.setLotteryPolicy(lotteryPolicy);
-        e.deactivateEvent();
-        eventRepository.save(e);
-        Result<Boolean> res = purchasePolicyService.enrollInLottery("user1", e1.getEventID());
-        assertFalse(res.isSuccess());
-        Event e2 = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertFalse(e2.getLotteryPolicy().getParticipants().contains("testuser"));
-    }
-
-    @Test
-    public void enrollLotteryPolicy_FailPastEnrollDate() throws InterruptedException {
-        Event e = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        LotteryPolicy lotteryPolicy = new LotteryPolicy(0, "Lottery", 50, now.plusSeconds(1));
-        e.setLotteryPolicy(lotteryPolicy);
-        eventRepository.save(e);
-        Thread.sleep(1000);
-        Result<Boolean> res = purchasePolicyService.enrollInLottery("user1", e1.getEventID());
-        assertFalse(res.isSuccess());
-        Event e2 = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertFalse(e2.getLotteryPolicy().getParticipants().contains("testuser"));
-    }
-
-    @Test
-    public void enrollLotteryPolicy_FailEnrollTwice() {
-        Event e = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        LotteryPolicy lotteryPolicy = new LotteryPolicy(0, "Lottery", 50, now.plusDays(5));
-        e.setLotteryPolicy(lotteryPolicy);
-        eventRepository.save(e);
-        Result<Boolean> res = purchasePolicyService.enrollInLottery("user1", e1.getEventID());
-        assertTrue(res.isSuccess());
-        Event e2 = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertTrue(e2.getLotteryPolicy().getParticipants().contains("testuser"));
-        Result<Boolean> res2 = purchasePolicyService.enrollInLottery("user1", e1.getEventID());
-        assertFalse(res2.isSuccess());
-        Event e3 = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertTrue(e3.getLotteryPolicy().getParticipants().contains("testuser"));
-    }
-
-    @Test
-    public void enrollLotteryPolicy_TwoThreadsBothSucceeds() throws InterruptedException {
-        Event eve = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        LotteryPolicy lotteryPolicy = new LotteryPolicy(0, "Lottery", 50, now.plusDays(5));
-        eve.setLotteryPolicy(lotteryPolicy);
-        eventRepository.save(eve);
-
-        CountDownLatch readyLatch = new CountDownLatch(2);
-        CountDownLatch startLatch = new CountDownLatch(1);
-
-        AtomicReference<Result<Boolean>> result1 = new AtomicReference<>();
-        AtomicReference<Result<Boolean>> result2 = new AtomicReference<>();
-
-        Runnable createPolicyTask1 = () -> {
-            try {
-                readyLatch.countDown();
-                startLatch.await();
-                result1.set(purchasePolicyService.enrollInLottery("user1", e1.getEventID()));
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-            }
-        };
-
-        Runnable createPolicyTask2 = () -> {
-            try {
-                readyLatch.countDown();
-                startLatch.await();
-                result2.set(purchasePolicyService.enrollInLottery("user2", e1.getEventID()));
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-            }
-        };
-
-        Thread thread1 = new Thread(createPolicyTask1);
-        Thread thread2 = new Thread(createPolicyTask2);
-
-        thread1.start();
-        thread2.start();
-
-        readyLatch.await();
-        startLatch.countDown();
-
-        thread1.join();
-        thread2.join();
-
-        assertTrue(result1.get().isSuccess(), result1.get().getError());
-        assertTrue(result2.get().isSuccess(), result2.get().getError());
-
-        Event e3 = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertTrue(e3.getLotteryPolicy().getParticipants().contains("testuser"));
-        assertTrue(e3.getLotteryPolicy().getParticipants().contains("testuser2"));
-
-        int successCount = 0;
-        if (result1.get() != null && result1.get().isSuccess()) {
-            successCount++;
-        }
-        if (result2.get() != null && result2.get().isSuccess()) {
-            successCount++;
-        }
-
-        assertEquals(2, successCount);
-
-        Event e = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertDoesNotThrow(() -> e.getLotteryPolicy());
-    }
-
-    @Test
-    public void createLotteryPolicy_TwoThreadsOneSucceeds() throws InterruptedException {
-        Event eve = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        LotteryPolicy lotteryPolicy = new LotteryPolicy(0, "Lottery", 50, now.plusDays(5));
-        eve.setLotteryPolicy(lotteryPolicy);
-        eventRepository.save(eve);
-
-        CountDownLatch readyLatch = new CountDownLatch(2);
-        CountDownLatch startLatch = new CountDownLatch(1);
-
-        AtomicReference<Result<Boolean>> result1 = new AtomicReference<>();
-        AtomicReference<Result<Boolean>> result2 = new AtomicReference<>();
-
-        Runnable createPolicyTask1 = () -> {
-            try {
-                readyLatch.countDown();
-                startLatch.await();
-                result1.set(purchasePolicyService.enrollInLottery("user1", e1.getEventID()));
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-            }
-        };
-
-        Runnable createPolicyTask2 = () -> {
-            try {
-                readyLatch.countDown();
-                startLatch.await();
-                result2.set(purchasePolicyService.enrollInLottery("user1", e1.getEventID()));
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-            }
-        };
-
-        Thread thread1 = new Thread(createPolicyTask1);
-        Thread thread2 = new Thread(createPolicyTask2);
-
-        thread1.start();
-        thread2.start();
-
-        readyLatch.await();
-        startLatch.countDown();
-
-        thread1.join();
-        thread2.join();
-
-        int successCount = 0;
-        if (result1.get() != null && result1.get().isSuccess()) {
-            successCount++;
-        }
-        if (result2.get() != null && result2.get().isSuccess()) {
-            successCount++;
-        }
-
-        assertTrue(successCount == 1);
-
-        Event e = eventRepository.findByID(Integer.toString(e1.getEventID()));
-        assertDoesNotThrow(() -> e.getLotteryPolicy());
     }
 
     @Test
